Implement different modes for url handling (#15115).
The url mode is optional and can be one of "absolute", "relative" and
"mixed".

Absolute mode can be used for compatibility with sass-lang whereas the
(default) mixed mode is compatible with previous versions of Vaadin Sass
Compiler. In mixed mode url(...) is interpreted as relative only for
simple properties.

In absolute mode a url appears in the compiled css in the same form as
in the source scss. In relative mode url(foo) appearing in imported
style sheet bar/baz.scss is output as url(bar/foo).

Change-Id: I98863871584680f372dd8f01624e16394425642f